pyproject.toml
[project]
name = "colectivos-ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.24.1",
    "python-dotenv>=1.0.1",
    "websocket-client>=1.8.0",
]
